Add partial support for formatting incoming HTML
Enable the html_formatting var to use this. Currently supports bold,
italics, and underlines with weechat attributes. Code, strikethroughs,
quotes, and breaks/rules are supported with plain text decorating them.

It would be reasonable to add support for colors and lists in the
future. Unfortunately, some things will never be possible with the
limitations of Weechat, but there may be ways to approximate them better.
All unsupported HTML tags are simply stripped from the message.

This mostly fixes #60.